sveltekit/src/lib/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-auth>=2.40.3",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
]
